+++ src/py_dss_interface/models/ActiveClass/ActiveClassI.py
@@ -18,20 +18,20 @@ class ActiveClassI(Base):
 
     The properties (parameter) are integer numbers and are described as follows."""
 
-    def first(self) -> int:
+    def _first(self) -> int:
         """Sets first element in the active class to be the active DSS object. If object is a CktElement,
         ActiveCktElement also points to this element. Returns 0 if none."""
         return self.dss_obj.ActiveClassI(0, 0)
 
-    def next(self) -> int:
+    def _next(self) -> int:
         """Sets next element in the active class to be the active DSS object.
         If object is a CktElement, ActiveCktElement also points to this element. Returns 0 if none."""
         return self.dss_obj.ActiveClassI(1, 0)
 
-    def num_elements(self) -> int:
+    def _num_elements(self) -> int:
         """Gets the number of elements in this class. Same as Count Property."""
         return self.dss_obj.ActiveClassI(2, 0)
 
-    def count(self) -> int:
+    def _count(self) -> int:
         """Gets the number of elements in this class. Same as NumElements Property."""
         return self.dss_obj.ActiveClassI(3, 0)