+++ src/py_dss_interface/models/Example/ExampleBase.py
@@ -6,7 +6,7 @@
 import json
 import pathlib
 
-from py_dss_interface import DSSDLL as DSS
+from py_dss_interface import DSS as DSS
 
 
 class ExampleBase: