+++ CHANGELOG.rst
@@ -45,7 +45,7 @@ Changelog
 0.0.4 (2020-08-17)
 ------------------
 
-* DSSDLL class can receive OpenDSS folder in order to use the user own OpenDSS.
+* DSS class can receive OpenDSS folder in order to use the user own OpenDSS.
 
 0.0.1 (2020-06-12)
 ------------------